requirements.txt

fastapi==0.65.1
gunicorn==19.9.0
pandas==1.0.0
turicreate==6.4
uvicorn==0.13.4
numpy==1.20.2
pydantic==1.8.2
python-Levenshtein==0.12.2
pykakasi==2.1.1
googletrans==4.0.0rc1